+++ cli/internal_testsuites/golang/testsuite/network_soft_partition_test/network_soft_partition_test.go
@@ -206,9 +206,9 @@ func repartitionNetwork(
 	return nil
 }
 
-func getExampleServiceConfigSupplier() func(ipAddr string, sharedDirectory *services.SharedPath) (*services.ContainerConfig, error) {
+func getExampleServiceConfigSupplier() func(ipAddr string) (*services.ContainerConfig, error) {
 	portSpec := services.NewPortSpec(exampleServicePortNumInsideNetwork, services.PortProtocol_TCP)
-	containerConfigSupplier := func(ipAddr string, sharedDirectory *services.SharedPath) (*services.ContainerConfig, error) {
+	containerConfigSupplier := func(ipAddr string) (*services.ContainerConfig, error) {
 		containerConfig := services.NewContainerConfigBuilder(
 			dockerGettingStartedImage,
 		).WithUsedPorts(
@@ -219,8 +219,8 @@ func getExampleServiceConfigSupplier() func(ipAddr string, sharedDirectory *serv
 	return containerConfigSupplier
 }
 
-func getTestServiceContainerConfigSupplier() func(ipAddr string, sharedDirectory *services.SharedPath) (*services.ContainerConfig, error) {
-	containerConfigSupplier := func(ipAddr string, sharedDirectory *services.SharedPath) (*services.ContainerConfig, error) {
+func getTestServiceContainerConfigSupplier() func(ipAddr string) (*services.ContainerConfig, error) {
+	containerConfigSupplier := func(ipAddr string) (*services.ContainerConfig, error) {
 
 		// We sleep because the only function of this container is to test Docker executing a command while it's running
 		// NOTE: We could just as easily combine this into a single array (rather than splitting between ENTRYPOINT and CMD